fixed config load for floats

+++ src/constant/general.py
@@ -2,4 +2,4 @@ from util.config import load_config
 
 _config = load_config()["general"]
 
-SLEEP_TIME: int = int(_config["sleep_time"])
+SLEEP_TIME: float = float(_config["sleep_time"])
